Move check-call into check app
